nsuse() fixed to support more forms of import
Previously, some cases were not supported or caused errors.

The current version supports also forms described in
https://secure.php.net/manual/en/language.namespaces.importing.php
(multiple aliases per one line)
and does not cause errors when the parsed file was saved
with incomplete import line (such lines is ignored).

Yes, regular expressions are slower than string functions,
but it is the cost of relative simplicity and readibility.

Because of the need to manually kill the phpcd processes
with mentioned errors I decided to push it without making PR.