fix pseudo property detect error
If a pseudo is defined in the class's parent and the @property read a
relative class name (with out the absolute namespace), phpcd will use
the current class's namespace to resolve the absolute classname for the
property type hint, which will be not the result we want.

So this path use the ReflectionProperty::getDeclaringClass method to
fetch the right sourcefile and then use the namespace there to resolve
the right pseudo property's absolute classname. It just works :)